fix: redirect respects rewrites
fixes #5300
