+++ examples/react/basic-ssr-file-based/src/entry-server.tsx
@@ -18,7 +18,7 @@ export async function render({
   res: express.Response
 }) {
   // Convert the express request to a fetch request
-  const url = new URL(req.originalUrl || req.url, 'https://localhost:3000').href
+  const url = new URL(req.originalUrl || req.url, 'http://localhost:3000').href
 
   const request = new Request(url, {
     method: req.method,
