+++ packages/router-core/src/ssr/createRequestHandler.ts
@@ -1,6 +1,6 @@
 import { createMemoryHistory } from '@tanstack/history'
 import { mergeHeaders } from './headers'
-import { attachRouterServerSsrUtils } from './ssr-server'
+import { attachRouterServerSsrUtils, getOrigin } from './ssr-server'
 import type { HandlerCallback } from './handlerCallback'
 import type { AnyRouter } from '../router'
 import type { Manifest } from '../manifest'
@@ -27,7 +27,7 @@ export function createRequestHandler<TRouter extends AnyRouter>({
     })
 
     const url = new URL(request.url, 'http://localhost')
-
+    const origin = getOrigin(request)
     const href = url.href.replace(url.origin, '')
 
     // Create a history for the router
@@ -38,6 +38,7 @@ export function createRequestHandler<TRouter extends AnyRouter>({
     // Update the router with the history and context
     router.update({
       history,
+      origin: router.options.origin ?? origin,
     })
 
     await router.load()
