+++ packages/router-core/src/ssr/server.ts
@@ -7,7 +7,7 @@ export {
   transformStreamWithRouter,
   transformReadableStreamWithRouter,
 } from './transformStreamWithRouter'
-export { attachRouterServerSsrUtils } from './ssr-server'
+export { attachRouterServerSsrUtils, getOrigin } from './ssr-server'
 
 // declare module '../router' {
 //   export interface RegisterSsr {
